Make changes so that ajax options are applied
So when AJAX options are passed into this function, it would pass them
into `$.ajax`. Because of the promise it creates here, there would
possibly an extra unhandled promise for users who wished to maintain
compatibility with the default AJAX interface.

This fixes that also with it, it closes toggl/webapp#1633.